requirements.txt
asgiref==3.7.2
Django==5.0.2
psycopg2-binary==2.9.9
sqlparse==0.4.4
pillow==10.2.0
whitenoise==6.6.0
environs[django]==10.3.0
gunicorn==21.2.0
django-debug-toolbar==4.3.0